src/serveur/vote/requirements.txt
django
djangorestframework
neo4j